Improve temp sensor usage on Windows
Only initialize COM once, load and save the temperature sensor when
creating the Device struct, and make sure the sensor we load is actually
associated with the correct USB device.

+++ cli/src/device.rs
@@ -63,7 +63,7 @@ impl Device {
         self.d.send_report(report)
     }
 
-    pub fn read_temperature(&self) -> Result<i16, Error> {
+    pub fn read_temperature(&self) -> Result<f64, Error> {
         self.d.read_temperature()
     }
 }
